title component made for cleanup
